[BE] add develop script

+++ admin/index.js
@@ -2,8 +2,10 @@ const fse = require('fs-extra')
 const execa = require('execa')
 const path = require('path')
 
-const calcEnvVariables = (configs) => {
+const calcEnvVariables = (configs, isDevelopment) => {
   return [
+    '--env',
+    `ENV=${isDevelopment ? 'development' : 'production'}`,
     '--env',
     `ADMIN_PATH=${configs.admin.adminPath}`,
     '--env',
@@ -11,21 +13,24 @@ const calcEnvVariables = (configs) => {
   ]
 }
 
-const buildAdmin = async ({ cwd, configs }) => {
+const buildAdmin = async ({ cwd, configs }, isDevelopment) => {
   try {
     console.log(cwd, configs)
     const buildDir = path.resolve(cwd, 'build')
-    // if (!fse.pathExistsSync(buildDir)) {
-    await fse.emptyDir(path.resolve(__dirname, 'build'))
+    if (!fse.pathExistsSync(buildDir)) {
+      await fse.emptyDir(path.resolve(__dirname, 'build'))
 
-    console.log('Installing...')
-    await execa('yarn', ['install'], { cwd: __dirname })
-    
-    console.log('Building...')
-    await execa('yarn', ['build', ...calcEnvVariables(configs)], { cwd: __dirname })
+      console.log('Installing...')
+      await execa('yarn', ['install'], { cwd: __dirname })
 
-    await fse.copy(path.join(__dirname, 'build'), path.join(buildDir))
-    // }
+      console.log('Building...')
+      await execa('yarn', ['build', ...calcEnvVariables(configs, isDevelopment)], {
+        cwd: __dirname,
+      })
+
+      await fse.copy(path.join(__dirname, 'build'), path.join(buildDir))
+    }
+
     return console.log('Done!')
   } catch (error) {
     return console.error(error)
